Carriers can wait
Carriers can be told to goToIdle in which case they return to a
waitingLocation and stay there, also did some small fixes. Doesn't
remove carriers from tick list in board when idle, should be done
eventually.  Also not sure if it's the best implementation if waiting,
added a TODO to consider this.