Item passing works
There is a crude way to test item passing, and it seems to work. Managed
to make things crash at some point when creating extra roads, but might
have been a problem of making an order to a node that's not connected by
carriers. That still makes it crash, so some stuff still needs to be
fixed